Fix missing TagPrefix in Fluentd logger configuration
The Megaphone Fluentd container is configured to match streams
using a prefix: e.g. megaphone.work-updates

That prefix is provided as part of the MessagePack format of
the messages that are sent by the Fluentd loggers to Fluentd,
and is part of the configuration of the logger.

If it is missing, the messages are not matched properly and get
published to the topic-mismatches topic, the catch-all Megaphone topic.